requirements.txt
mwclient==0.10.0
python-twitter==3.5
